requirements.txt
anyio==3.2.1
argh==0.26.2
argon2-cffi==20.1.0
async_generator==1.10
attrs==21.2.0
babel==2.9.1
backcall==0.2.0
bleach==3.3.1
brotlipy==0.7.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
configparser==5.0.2
cryptography==3.4.7
cycler==0.10.0
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
docker-pycreds==0.4.0
einops==0.3.0
entrypoints==0.3
gitdb==4.0.10
gitpython==3.1.32
gql==0.1.0
graphql-core==3.1.5
idna==2.10
importlib-metadata==4.6.1
ipdb==0.13.9
ipykernel==6.0.2
ipython==7.25.0
ipython_genutils==0.2.0
jedi==0.18.0
jinja2==3.0.1
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter_client==6.1.12
jupyter_core==4.7.1
jupyter_server==1.9.0
jupyterlab==3.0.16
jupyterlab_pygments==0.1.2
jupyterlab_server==2.6.1
kiwisolver==1.3.1
liac-arff==2.5.0
markupsafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
mkl-service==2.4.0
mkl_fft==1.3.1
mkl_random==1.2.2
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.22.3
nvidia-ml-py==12.535.77
olefile==0.46
openml==0.11.0
packaging==21.0
pandas==1.2.4
pandocfilters==1.4.2
parso==0.8.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==8.2.0
prometheus_client==0.11.0
promise==2.3
prompt-toolkit==3.0.19
protobuf==3.17.2
psutil==5.8.0
ptyprocess==0.7.0
pycparser==2.20
pygments==2.9.0
pyopenssl==20.0.1
pyparsing==2.4.7
# PyQt5==5.15.9
pyrsistent==0.17.3
pysocks==1.7.1
python-dateutil==2.8.1
wget==3.2
torch==1.8.1
pytz==2021.1
pyyaml==5.4.1
pyzmq==22.1.0
requests==2.25.1
requests-unixsocket==0.2.0
scikit-learn==0.23.2
scipy==1.6.2
seaborn==0.11.0
send2trash==1.7.1
sentry-sdk==1.1.0
setuptools==49.6.0
shortuuid==1.0.1
sip==6.7.9
six==1.16.0
smmap==3.0.5
sniffio==1.2.0
subprocess32==3.5.4
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.1.0
torchvision==0.9.1
tornado==6.1
traitlets==5.0.5
typing_extensions==3.7.4.3
urllib3==1.26.5
wandb==0.10.31
watchdog==0.10.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
wheel==0.36.2
xgboost==1.4.0
xmltodict==0.12.0
zipp==3.5.0
